+++ app/dashboard/advanced-settings/at-commands/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import CommonATCommandsComponent from "@/components/at-terminal/common-commands";
+
+const CommonATCommandsPage = () => {
+  return <CommonATCommandsComponent />;
+};
+
+export default CommonATCommandsPage;